first version

+++ common/src/main/kotlin/cz/lukynka/cloudsettingssync/common/MusicAndSounds.kt
@@ -4,16 +4,16 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class MusicAndSounds(
-    val master: Int,
-    val music: Int,
-    val weather: Int,
-    val hostileCreatures: Int,
-    val players: Int,
-    val voice: Int,
-    val jukebox: Int,
-    val blocks: Int,
-    val friendlyCreatures: Int,
-    val ambient: Int,
+    val master: Float,
+    val music: Float,
+    val weather: Float,
+    val hostileCreatures: Float,
+    val players: Float,
+    val voice: Float,
+    val jukebox: Float,
+    val blocks: Float,
+    val friendlyCreatures: Float,
+    val ambient: Float,
     val showSubtitles: Boolean,
     val directionalAudio: Boolean
 )
